+++ geode-core/src/distributedTest/java/org/apache/geode/internal/cache/tier/sockets/DurableResponseMatrixDUnitTest.java
@@ -79,7 +79,7 @@ public class DurableResponseMatrixDUnitTest extends JUnit4DistributedTestCase {
     final Host host = Host.getHost(0);
     server1 = host.getVM(0);
     // start servers first
-    PORT1 = ((Integer) server1.invoke(() -> DurableResponseMatrixDUnitTest.createServerCache()));
+    PORT1 = server1.invoke(() -> DurableResponseMatrixDUnitTest.createServerCache());
     createCacheClient(NetworkUtils.getServerHostName(server1.getHost()));
     // Disconnecting the client can cause this
     IgnoredException.addIgnoredException("Connection reset||Unexpected IOException");
