+++ geode-core/src/distributedTest/java/org/apache/geode/internal/cache/tier/sockets/InterestListRecoveryDUnitTest.java
@@ -95,9 +95,9 @@ public class InterestListRecoveryDUnitTest extends JUnit4DistributedTestCase {
     server1 = host.getVM(0);
     server2 = host.getVM(1);
     // start servers first
-    PORT1 = ((Integer) server1.invoke(() -> InterestListRecoveryDUnitTest.createServerCache()))
+    PORT1 = server1.invoke(() -> InterestListRecoveryDUnitTest.createServerCache())
         .intValue();
-    PORT2 = ((Integer) server2.invoke(() -> InterestListRecoveryDUnitTest.createServerCache()))
+    PORT2 = server2.invoke(() -> InterestListRecoveryDUnitTest.createServerCache())
         .intValue();
 
     org.apache.geode.test.dunit.LogWriterUtils.getLogWriter()
