+++ geode-core/src/distributedTest/java/org/apache/geode/internal/cache/wan/misc/CommonAsyncEventQueueDUnitTest.java
@@ -77,7 +77,7 @@ public class CommonAsyncEventQueueDUnitTest extends AsyncEventQueueTestBase {
   public void testSameSenderWithNonColocatedRegions() throws Exception {
     IgnoredException.addIgnoredException("cannot have the same parallel async");
     Integer lnPort =
-        (Integer) vm0.invoke(() -> AsyncEventQueueTestBase.createFirstLocatorWithDSId(1));
+        vm0.invoke(() -> AsyncEventQueueTestBase.createFirstLocatorWithDSId(1));
     vm1.invoke(() -> AsyncEventQueueTestBase.createCache(lnPort));
     vm1.invoke(() -> AsyncEventQueueTestBase.createAsyncEventQueue("ln", true, 100, 100, false,
         false, null, false));
@@ -104,7 +104,7 @@ public class CommonAsyncEventQueueDUnitTest extends AsyncEventQueueTestBase {
     String regionName = getTestMethodName() + "_PR1";
 
     Integer lnPort =
-        (Integer) vm0.invoke(() -> AsyncEventQueueTestBase.createFirstLocatorWithDSId(1));
+        vm0.invoke(() -> AsyncEventQueueTestBase.createFirstLocatorWithDSId(1));
     vm1.invoke(() -> AsyncEventQueueTestBase.createCache(lnPort));
 
     vm1.invoke(() -> {
